+++ src/main/java/liquibase/sqlgenerator/ext/DropViewGeneratorHanaDB.java
@@ -23,6 +23,10 @@ public class DropViewGeneratorHanaDB extends DropViewGenerator {
 
     @Override
     public Sql[] generateSql(DropViewStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+        if (!supports(statement, database)) {
+            return sqlGeneratorChain.generateSql(statement, database);
+        }
+
         return new Sql[] {
                 new UnparsedSql(
                         "DROP VIEW " + database.escapeViewName(statement.getCatalogName(), statement.getSchemaName(), statement.getViewName()),
